avoid editMap to inherit sync only used in Visu
